move stuff to a manager outside the hook + more small cleanups
