+++ src/main/java/at/hannibal2/skyhanni/features/garden/SensitivityReducer.kt
@@ -7,7 +7,8 @@ import at.hannibal2.skyhanni.config.commands.CommandCategory
 import at.hannibal2.skyhanni.config.commands.CommandRegistrationEvent
 import at.hannibal2.skyhanni.config.features.garden.SensitivityReducerConfig
 import at.hannibal2.skyhanni.events.GuiRenderEvent
-import at.hannibal2.skyhanni.mixins.hooks.MouseSensitivityHook
+import at.hannibal2.skyhanni.features.garden.sensitivity.MouseSensitivityManager
+import at.hannibal2.skyhanni.features.garden.sensitivity.MouseSensitivityManager.SensitivityState
 import at.hannibal2.skyhanni.skyhannimodule.SkyHanniModule
 import at.hannibal2.skyhanni.utils.ChatUtils
 import at.hannibal2.skyhanni.utils.KeyboardManager.isKeyHeld
@@ -24,9 +25,9 @@ object SensitivityReducer {
 
     private var shouldBeActive = false
     private val isActive get() = isAutoActive || isManualActive
-    private val isAutoActive get() = MouseSensitivityHook.state == MouseSensitivityHook.SensitivityState.AUTO_REDUCED
-    private val isManualActive get() = MouseSensitivityHook.state == MouseSensitivityHook.SensitivityState.MANUAL_REDUCED
-    private val isMouseLocked get() = MouseSensitivityHook.state == MouseSensitivityHook.SensitivityState.LOCKED
+    private val isAutoActive get() = SensitivityState.AUTO_REDUCED.isActive()
+    private val isManualActive get() = SensitivityState.MANUAL_REDUCED.isActive()
+    private val isMouseLocked get() = SensitivityState.LOCKED.isActive()
 
     @HandleEvent
     fun onTick() {
@@ -60,10 +61,10 @@ object SensitivityReducer {
 
         if (!isActive) {
             shouldBeActive = true
-            MouseSensitivityHook.setState(MouseSensitivityHook.SensitivityState.AUTO_REDUCED)
+            MouseSensitivityManager.state = SensitivityState.AUTO_REDUCED
         } else {
             shouldBeActive = false
-            MouseSensitivityHook.setState(MouseSensitivityHook.SensitivityState.UNCHANGED)
+            MouseSensitivityManager.state = SensitivityState.UNCHANGED
         }
     }
 
@@ -87,21 +88,21 @@ object SensitivityReducer {
         if (config.onGround.get() && !onGround) return
         if (!isActive) {
             shouldBeActive = true
-            MouseSensitivityHook.setState(MouseSensitivityHook.SensitivityState.AUTO_REDUCED)
+            MouseSensitivityManager.state = SensitivityState.AUTO_REDUCED
         } else {
             shouldBeActive = false
-            MouseSensitivityHook.setState(MouseSensitivityHook.SensitivityState.UNCHANGED)
+            MouseSensitivityManager.state = SensitivityState.UNCHANGED
         }
     }
 
     private fun manualToggle() {
         if (!isActive) {
             shouldBeActive = true
-            MouseSensitivityHook.setState(MouseSensitivityHook.SensitivityState.MANUAL_REDUCED)
+            MouseSensitivityManager.state = SensitivityState.MANUAL_REDUCED
             ChatUtils.chat("§bMouse sensitivity is now lowered. Type /shsensreduce to restore your sensitivity.")
         } else {
             shouldBeActive = false
-            MouseSensitivityHook.setState(MouseSensitivityHook.SensitivityState.UNCHANGED)
+            MouseSensitivityManager.state = SensitivityState.UNCHANGED
             ChatUtils.chat("§bMouse sensitivity is now restored.")
         }
     }
